Version Beta 0.2.3! / Версия Beta 0.2.3! - sales system fixes: -- the number of resources and money are calculated in long; -- when displaying large numbers, scientific transliteration is used; --------------------------------------------------------------------- - исправления системы продажи: -- число ресурсов и деньги вычисляются в long; -- при выводе больших чисел используется научная транслитерация;

+++ Assets/Scripts/Cargos/Cargo.cs
@@ -23,7 +23,7 @@ namespace PlanetsColony
             return 0f;
         }
 
-        public override uint GetValue()
+        public override ulong GetValue()
         {
             return value;
         }
